Added more installation info to README - Moved Finish button from reservation to table
