+++ README.md
@@ -87,15 +87,15 @@ Clicking cancel will take the user back to the previous page.
 
 ### Install
 
-```sh
-npm install
-```
+1. Run `cp ./back-end/.env.sample ./back-end/.env`.
+1. Update the `./back-end/.env` file with the connection URL's to your PostgreSQL database instance.
+1. Run `cp ./front-end/.env.sample ./front-end/.env`.
+1. You should not need to make changes to the `./front-end/.env` file unless you want to connect to a backend at a location other than `http://localhost:5001`.
+1. Run `npm install` to install project dependencies.
 
 ### Usage
 
-```sh
-npm start
-```
+`npm run start:dev`
 
 ### Run tests
